refact admin ui
