+++ web/gds-admin-ui/src/pages/app/details/BasicDetails/dashboard/__tests__/StatisticCard.spec.js
@@ -1,0 +1,6 @@
+
+describe('<StatisticCard />', () => {
+
+    it.todo('should render props correctly')
+    it.todo('should render count as 0')
+})
