+++ web/gds-admin-ui/src/pages/app/details/BasicDetails/dashboard/index.js
