+++ web/gds-admin-ui/src/pages/app/details/CertificateDetails/IdentityCertificateDropDown.js
@@ -1,0 +1,32 @@
+import { Dropdown } from 'react-bootstrap';
+import PropTypes from 'prop-types'
+
+const IdentityCertificateDropDown = ({ handleCopySignatureClick, handleCopySerialNumberClick }) => {
+
+    return (
+        <Dropdown className="float-end" align="end">
+            <Dropdown.Toggle
+                data-testid="certificate-details-3-dots"
+                variant="link"
+                tag="a"
+                className="card-drop arrow-none cursor-pointer p-0 shadow-none">
+                <i className="dripicons-dots-3"></i>
+            </Dropdown.Toggle>
+            <Dropdown.Menu>
+                <Dropdown.Item data-testid="copy-signature" onClick={handleCopySignatureClick}>
+                    <i className="mdi mdi-content-copy me-1"></i>Copy signature
+                </Dropdown.Item>
+                <Dropdown.Item data-testid="copy-serial-number" onClick={handleCopySerialNumberClick}>
+                    <i className="mdi mdi-content-copy me-1"></i>Copy serial number
+                </Dropdown.Item>
+            </Dropdown.Menu>
+        </Dropdown>
+    )
+}
+
+IdentityCertificateDropDown.propTypes = {
+    handleCopySignatureClick: PropTypes.func.isRequired,
+    handleCopySerialNumberClick: PropTypes.func.isRequired,
+}
+
+export default IdentityCertificateDropDown
